requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.24
pymupdf>=1.23.9
pdf2image>=1.17.0
pillow>=10.2.0
python-docx>=1.1.0
pydantic>=2.7.1
rich>=13.7.1
tqdm>=4.66.4
gradio>=4.31.8
ollama>=0.1.9
sentence-transformers>=2.7.0
jinja2>=3.1.3
numpy>=1.26.4
scikit-learn>=1.4.2
psutil>=5.9.8
pynvml>=11.5.0
